refactor(db): extract checkpoint retention ownership
